+++ src/main/client/controllers/index.ts
@@ -1,6 +1,7 @@
 import * as alt from 'alt-client';
 
 import './blip.js';
+import './d2dTextLabel.js';
 import './interaction.js';
 import './marker.js';
 import './object.js';