+++ src/main/client/virtualEntities/index.ts
@@ -1,5 +1,6 @@
 import * as alt from 'alt-client';
 
+import './d2dTextLabel.js';
 import './marker.js';
 import './progressbar.js';
 import './textlabel.js';