+++ src/main/server/vehicle/index.ts
@@ -2,6 +2,7 @@ import * as alt from 'alt-server';
 import { useRebar } from '../index.js';
 import { Vehicle, WheelState } from '../../shared/types/vehicle.js';
 import * as Utility from '@Shared/utility/index.js';
+import { Events } from '../../shared/events/index.js';
 
 const Rebar = useRebar();
 const db = Rebar.database.useDatabase();
@@ -82,12 +83,12 @@ export function useVehicle(vehicle: alt.Vehicle) {
         }
 
         // Synchronize wheelColor
-        if(typeof document.wheelColor !== 'undefined') {
+        if (typeof document.wheelColor !== 'undefined') {
             vehicle.wheelColor = document.wheelColor;
         }
 
         // Synchronize pearlColor
-        if(typeof document.pearlColor !== 'undefined') {
+        if (typeof document.pearlColor !== 'undefined') {
             vehicle.pearlColor = document.pearlColor;
         }
 
@@ -496,6 +497,20 @@ export function useVehicle(vehicle: alt.Vehicle) {
         return Utility.vehicleHashes.getNameFromHash(vehicle.model);
     }
 
+    /**
+     * Set the RPM for the vehicle
+     *
+     * @param {number} value
+     * @return
+     */
+    function setRpm(value: number) {
+        if (!vehicle.driver) {
+            return;
+        }
+
+        alt.emitClient(vehicle.driver, Events.vehicle.set.rpm, value);
+    }
+
     return {
         apply,
         bind,
@@ -510,6 +525,7 @@ export function useVehicle(vehicle: alt.Vehicle) {
         },
         repair,
         save,
+        setRpm,
         sync,
         toggleDoor,
         toggleDoorAsPlayer,