+++ src/main/shared/events/index.ts
@@ -30,6 +30,10 @@ export const Events = {
             create: 'textlabel:create',
             destroy: 'textlabel:destroy',
         },
+        dxgilabel: {
+            create: 'dxgilabel:create',
+            destroy: 'dxgilabel:destroy',
+        },
     },
     events: {
         clientEvents: {
@@ -98,6 +102,11 @@ export const Events = {
             update: 'systems:keybinds:update',
             invoke: 'systems:keybinds:invoke',
         },
+        keypress: {
+            update: 'systems:keypress:update',
+            invokeUp: 'systems:keypress:invoke:keyup',
+            invokeDown: 'systems:keypress:invoke:keydown',
+        },
         messenger: {
             process: 'systems:messenger:process',
             send: 'systems:messenger:send',
@@ -125,6 +134,11 @@ export const Events = {
         },
         transmitter: {
             execute: 'systems:transmitter:execute',
+        },
+    },
+    vehicle: {
+        set: {
+            rpm: 'vehicle:set:rpm',
         },
     },
     view: {