+++ docs/changelog.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## Version 40
+
+### Code Changes
+
+-   Added new dlc clothing maximums for 2024 DLC
+-   Added `keypress` api that lets you bind functions to keyup/keydown from server-side
+-   Added `setRpm` to vehicle API
+-   Added new `fonts` resource, which may need to be imported in your `server.toml`; for custom fonts
+-   Lower distance for all Text Labels; max distance now capped at 20
+-   Failing readiness check on an API will now return the failing API's name
+-   Added `D2D` text labels with global and local support
+
+### Docs Changes
+
+-   Documented keypress
+-   Documented keypress in vehicle api
+-   Documented fonts folder
+
+---
+
 ## Version 39
 
 ### Code Changes